feat:#01 static init()method　not popular.but asynchronous operation used
